+++ src/main/java/com/mindera/finalproject/be/exception/course/CourseExceptions.java
@@ -1,4 +1,4 @@
-package com.mindera.finalproject.be.exception.CourseExceptions;
+package com.mindera.finalproject.be.exception.course;
 
 public class CourseExceptions extends Exception {
