+++ src/main/java/com/mindera/finalproject/be/exception/course/CourseNotFoundException.java
@@ -1,4 +1,4 @@
-package com.mindera.finalproject.be.exception.CourseExceptions;
+package com.mindera.finalproject.be.exception.course;
 
 public class CourseNotFoundException extends CourseExceptions{
