+++ src/main/java/com/mindera/finalproject/be/exception/student/studentAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.mindera.finalproject.be.exception.student;
+
+public class studentAlreadyExistsException extends studentException {
+    public studentAlreadyExistsException(String message) {
+        super(message);
+    }
+}
