+++ src/main/java/com/mindera/finalproject/be/exception/student/studentException.java
@@ -1,0 +1,7 @@
+package com.mindera.finalproject.be.exception.student;
+
+public class studentException extends Exception {
+    public studentException(String message) {
+        super(message);
+    }
+}
