+++ src/main/java/com/mindera/finalproject/be/exception/student/studentNotFoundException.java
@@ -1,0 +1,7 @@
+package com.mindera.finalproject.be.exception.student;
+
+public class studentNotFoundException extends studentException {
+    public studentNotFoundException(String message) {
+        super(message);
+    }
+}
